% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRings.R
\name{plot_circular_rings}
\alias{plot_circular_rings}
\title{Title: radial heatmap plotter}
\usage{
plot_circular_rings(
  data,
  ring_cols,
  annotation_col = "rs",
  ring_colors = NULL,
  ring_width = 0.5,
  title = "title",
  annotation_distance = 1,
  add_ticks = TRUE,
  tick_length = 0.1,
  fill_low = "white"
)
}
\arguments{
\item{data}{A data frame containing gene expression data. Each column represents
a different variable, and the row names should be gene IDs or SNPs.}

\item{ring_cols}{A vector of column names from the `data` frame to be plotted as rings.}

\item{annotation_col}{The column name for annotations (e.g., SNP IDs or gene names).}

\item{ring_colors}{A vector of colors to use for each ring.}

\item{ring_width}{The width of each ring.}

\item{title}{The title for the plot.}

\item{annotation_distance}{The distance from the center for placing annotations.}

\item{add_ticks}{Logical indicating whether to add ticks on the outer ring.}

\item{tick_length}{The length of the ticks.}

\item{fill_low}{The color used for the lowest values in the rings.}
}
\value{
A `ggplot` object representing the radial heatmap.
}
\description{
Description: This function generates a radial heatmap (circular rings)
expression data using ggplot2. It displays multiple layers (rings) representing
different gene expression metrics, along with optional annotations and ticks.
}
